chore: resolve PR comments
