+++ packages/core/lib/Diagnostics.ts
@@ -1,11 +1,26 @@
-export enum PluginEnum {
-  listener = "Listener",
-  crossover = "Crossover",
-  sampler = "Sampler",
-  searchAlgorithm = "Search algorithm",
-  terminationTrigger = "Termination trigger",
-  userInterface = "User-interface",
-}
+/*
+ * Copyright 2023-2023 Delft University of Technology and SynTest contributors
+ *
+ * This file is part of SynTest Framework - SynTest Core.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { PluginEnum } from "./plugin/PluginEnum";
+
+/**
+ * This file is meant to provide consistent error messages throughout the tool.
+ */
 
 export const pluginNotFound = (plugin: PluginEnum, name: string) =>
   `${plugin} plugin not found\nSpecified ${plugin} plugin: '${name}' not found in pluginManager.`;
