+++ packages/core/lib/search/metaheuristics/RandomSearch.ts
@@ -104,7 +104,7 @@ export class RandomSearchFactory<T extends Encoding>
     options: SearchAlgorithmOptions<T>
   ): SearchAlgorithm<T> {
     if (!options.eventManager) {
-      throw new Error("RandomSearch requires eventManager option.");
+      throw new Error(pluginRequiresOptions("RandomSearch", "eventManager"));
     }
     if (!options.encodingSampler) {
       throw new Error(pluginRequiresOptions("RandomSearch", "encodingSampler"));
